requirements.txt
streamlit>=1.20.0
transformers>=4.30.0
torch>=2.0.0
PyPDF2>=3.0.0
python-dotenv>=0.19.0
sentencepiece>=0.1.99
accelerate>=0.20.0
openai==0.28
python-dotenv>=0.19.0